requirements.txt
astroid==2.3.3
autopep8==1.4.4
dj-database-url==0.5.0
Django==2.2.7
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.4
wrapt==1.11.2
psycopg2-binary==2.7.5
django-heroku==0.3.1